+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Update using ```composer update```.
 * Added `{% nospace %}...{% endnospace %}` blocks to template syntax that will remove all whitespace between HTML tags.
 * Added `PaginationInterface::isValidPage()` method.
 * Mako is now using Whoops for displaying exception details.
+* Now possible register route constraints with the router.
 
 #### Changes